Refactor the loop into its own module.
