+++ .pylintrc
@@ -44,7 +44,8 @@ disable=wrong-import-position,
     too-many-arguments,
     redefined-builtin,
 	useless-suppression,
-	mixed-indentation
+	mixed-indentation,
+    useless-object-inheritance
 #	undefined-all-variable
